move static content loading to MineDashboard, remove debug
